feat: return stable cache digests via progress
We want to track the time saved of cached layers
over time but the public layer digest was not stable
from run to run.

This update adds a new stable digest to the cli
via the progress control.

The core addition was to notify progress during CacheMap
lookup.  There absolutely seem to be race conditions!

For some reason if I add any logic, such as filtering
for only those vertices that are completed, the progress
with a stable digest is _NOT_ sent to the client.

This makes me quesy.  More buildkit mysteries here.

Signed-off-by: Chris Goller <goller@gmail.com>